Checked against CMUCL 21b - no changes
I previously missed that CMUCL 21a already doesn't allow for slot :allocation other than :instance or :class anymore. This is still the case for CMUCL 21b. (It was not the case in previous versions of CMUCL, that is 20f or older.) I have now updated the documentation retroactively for 21a, and it stays the same for 21b. There is no fix for this issue in Closer to MOP.