Add implicit block and ignorable declarations in make-method-lambda.
According to the spec, every method body is enclosed in an implicit block
of the same name as the function.  Furthermore, for convenience, most
implementations declare all required arguments of a method as ignorable.
Prior to this commit, the implicit block and the ignorable declarations
were added as part of the expansion of closer-mop:defmethod, and before
supplying the lambda to make-method-lambda.  However, SBCL and ABCL do not
use the defmethod macro of closer-mop, but fall back to cl:defmethod.  As a
result, neither the implicit block, nor ignorable declarations were added
on SBCL and ABCL at all.  The reason why this bug was not detected earlier
is that this version of make-method-lambda is only used for custom generic
functions and methods.

This commit attempts to make the behavior of defmethod and
make-method-lambda uniform and correct across all implementation.  The
method body is now not at all manipulated by defmethod, but only by
make-method-lambda.